Move some classes out of Instrumented.java
Since this class is getting bigger and bigger.
