+++ subprojects/core/src/main/java/org/gradle/initialization/DefaultClassLoaderRegistry.java
@@ -18,7 +18,7 @@ package org.gradle.initialization;
 
 import org.gradle.api.internal.ClassPathRegistry;
 import org.gradle.internal.classloader.FilteringClassLoader;
-import org.gradle.internal.classpath.Instrumented.CallInterceptorRegistry;
+import org.gradle.internal.classpath.intercept.CallInterceptorRegistry;
 import org.gradle.internal.reflect.Instantiator;
 
 public class DefaultClassLoaderRegistry implements ClassLoaderRegistry {
