+++ subprojects/core/src/main/java/org/gradle/internal/classpath/Instrumented.java
@@ -22,14 +22,8 @@ import org.codehaus.groovy.runtime.ResourceGroovyMethods;
 import org.codehaus.groovy.runtime.callsite.CallSite;
 import org.codehaus.groovy.runtime.callsite.CallSiteArray;
 import org.codehaus.groovy.vmplugin.v8.IndyInterface;
-import org.gradle.api.NonNullApi;
 import org.gradle.internal.SystemProperties;
-import org.gradle.internal.classpath.GroovyCallInterceptorsProvider.ClassLoaderSourceGroovyCallInterceptorsProvider;
-import org.gradle.internal.classpath.JvmBytecodeInterceptorSet.ClassLoaderSourceJvmBytecodeInterceptorSet;
 import org.gradle.internal.classpath.intercept.CallInterceptor;
-import org.gradle.internal.classpath.intercept.CallInterceptorResolver;
-import org.gradle.internal.classpath.intercept.CallInterceptorsSet;
-import org.gradle.internal.classpath.intercept.CallSiteDecorator;
 import org.gradle.internal.classpath.intercept.ClassBoundCallInterceptor;
 import org.gradle.internal.classpath.intercept.InterceptScope;
 import org.gradle.internal.classpath.intercept.Invocation;
@@ -48,16 +42,14 @@ import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Properties;
-import java.util.WeakHashMap;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
-import static org.gradle.internal.classpath.Instrumented.CallInterceptorRegistry.getGroovyCallDecorator;
+import static org.gradle.internal.classpath.intercept.CallInterceptorRegistry.getGroovyCallDecorator;
 import static org.gradle.internal.classpath.MethodHandleUtils.findStaticOrThrowError;
 import static org.gradle.internal.classpath.MethodHandleUtils.lazyKotlinStaticDefaultHandle;
 
@@ -104,33 +96,6 @@ public class Instrumented {
         }
     };
 
-    @NonNullApi
-    public static class CallInterceptorRegistry {
-        private static final Map<ClassLoader, Boolean> LOADED_FROM_CLASSLOADERS = Collections.synchronizedMap(new WeakHashMap<>());
-        private static volatile CallSiteDecorator currentGroovyCallDecorator = new CallInterceptorsSet(GroovyCallInterceptorsProvider.DEFAULT);
-        private static volatile JvmBytecodeInterceptorSet currentJvmBytecodeInterceptors = JvmBytecodeInterceptorSet.DEFAULT;
-
-        public synchronized static void loadCallInterceptors(ClassLoader classLoader) {
-            if (LOADED_FROM_CLASSLOADERS.put(classLoader, true) != null) {
-                throw new RuntimeException("Cannot load interceptors twice for class loader: " + classLoader);
-            }
-
-            GroovyCallInterceptorsProvider classLoaderGroovyCallInterceptors = new ClassLoaderSourceGroovyCallInterceptorsProvider(classLoader);
-            GroovyCallInterceptorsProvider callInterceptors = GroovyCallInterceptorsProvider.DEFAULT.plus(classLoaderGroovyCallInterceptors);
-            currentGroovyCallDecorator = new CallInterceptorsSet(callInterceptors);
-            ClassLoaderSourceJvmBytecodeInterceptorSet classLoaderJvmBytecodeInterceptors = new ClassLoaderSourceJvmBytecodeInterceptorSet(classLoader);
-            currentJvmBytecodeInterceptors = JvmBytecodeInterceptorSet.DEFAULT.plus(classLoaderJvmBytecodeInterceptors);
-        }
-
-        public static CallSiteDecorator getGroovyCallDecorator() {
-            return currentGroovyCallDecorator;
-        }
-
-        public static JvmBytecodeInterceptorSet getJvmBytecodeInterceptors() {
-            return currentJvmBytecodeInterceptors;
-        }
-    }
-
     private static final AtomicReference<Listener> LISTENER = new AtomicReference<>(NO_OP);
 
     public static void setListener(Listener listener) {
@@ -163,44 +128,6 @@ public class Instrumented {
             new ProcessBuilderStartInterceptor(),
             new ProcessBuilderStartPipelineInterceptor()
         );
-    }
-
-    @NonNullApi
-    private static final class InterceptorResolverImpl implements CallInterceptorResolver {
-        @Nullable
-        @Override
-        public CallInterceptor resolveCallInterceptor(InterceptScope scope) {
-            CallSiteDecorator currentDecorator = getGroovyCallDecorator();
-            if (currentDecorator instanceof CallInterceptorResolver) {
-                return ((CallInterceptorResolver) currentDecorator).resolveCallInterceptor(scope);
-            }
-            return null;
-        }
-
-        @Override
-        public boolean isAwareOfCallSiteName(String name) {
-            CallSiteDecorator currentDecorator = getGroovyCallDecorator();
-            if (currentDecorator instanceof CallInterceptorResolver) {
-                return ((CallInterceptorResolver) currentDecorator).isAwareOfCallSiteName(name);
-            }
-            return false;
-        }
-    }
-
-    public static final CallInterceptorResolver INTERCEPTOR_RESOLVER = new InterceptorResolverImpl();
-
-    // TODO: We can support getting the interceptors in the instrumented code from different locations, not just `Instrumented.currentCallDecorator`,
-    //       so that replacing the call interceptors for tests would instead work by embedding a different location
-    //       than this one in the instrumented code. Doing that adds much more complexity in the instrumentation.
-    @NonNullApi
-    public static class GroovyCallInterceptorInternalTesting {
-        static CallSiteDecorator getCurrentGroovyCallSiteDecorator() {
-            return getGroovyCallDecorator();
-        }
-
-        static void setCurrentGroovyCallSiteDecorator(CallSiteDecorator interceptorsSet) {
-            CallInterceptorRegistry.currentGroovyCallDecorator = interceptorsSet;
-        }
     }
 
     // Called by generated code
